abstract format vice function
