+++ src/Components/Agenda/AgendaItemLegs/AgendaItemLegs.jsx
@@ -3,6 +3,7 @@ import { shortenString } from 'utilities';
 import { filter, take, takeRight } from 'lodash';
 import { format, isDate } from 'date-fns-v2';
 import FA from 'react-fontawesome';
+import { formatVice } from '../Constants';
 
 const AgendaItemLegs = props => {
   const {
@@ -26,18 +27,6 @@ const AgendaItemLegs = props => {
   const formatLang = (langArr = []) => langArr.map(lang => (
     `${lang.code} ${lang.spoken_proficiency}/${lang.reading_proficiency}`
   )).join(', ');
-  const formatVice = (viceObj) => {
-    const first = viceObj?.emp_first_name;
-    const last = viceObj?.emp_last_name;
-    const vice = (first || last)
-      ? `${first ? `${first} ` : ''}${last || ''}`
-      : '';
-    const vacancy = viceObj?.asgd_etd_ted_date && formatDate(viceObj.asgd_etd_ted_date);
-    if (vice || vacancy) {
-      return `${vice || ''}${(vice && vice !== 'Multiple Incumbents' && vacancy) ? ', ' : ''} ${vacancy || ''}`;
-    }
-    return '-';
-  };
 
   const getData = (key, helperFunc = () => {}) => (
     <>
