+++ src/Components/Agenda/Constants.js
@@ -33,3 +33,16 @@ export const statusRenaming = [
 /* eslint-enable import/prefer-default-export */
 
 export const dateTernary = date => date ? `${formatDate(date, 'MM/DD/YY')}` : '--/--/--';
+
+export const formatVice = (viceObj) => {
+  const first = viceObj?.emp_first_name;
+  const last = viceObj?.emp_last_name;
+  const vice = (first || last)
+    ? `${first ? `${first} ` : ''}${last || ''}`
+    : '';
+  const vacancy = viceObj?.asgd_etd_ted_date && formatDate(viceObj.asgd_etd_ted_date, 'MM/YY');
+  if (vice || vacancy) {
+    return `${vice || ''}${(vice && vice !== 'Multiple Incumbents' && vacancy) ? ', ' : ''} ${vacancy || ''}`;
+  }
+  return '-';
+};
